third_party/WebKit/LayoutTests/editing/execCommand/insertHTML.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>|</div>',
    'insertHTML foo<table style="display:inline-table"><tr><td>1</td><td>2</td></tr></table><b>bar</b>!',
    [
        '<div contenteditable>',
            'foo',
            '<table style="display:inline-table"><tbody>',
                '<tr><td>1</td><td>2</td></tr>',
            '</tbody></table>',
            '<b>bar</b>!|',
        '</div>',
    ].join('')),
    'insertHTML');
</script>
